docker/airflow/requirements.txt
apache-airflow==2.5.0
apache-airflow-providers-apache-spark==4.0.0
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-apache-kafka==1.0.0
psycopg2-binary==2.9.5
pyspark==3.3.0 